moved fragments to ui package
